Task-53706: Edit a poll before posting (#8)
After this change, we will be able to edit the created poll before posting it as long as the create poll drawer is not closed and the page is not refreshed. The label "you have created a poll" will be displayed in the main activity composer instead of "create a poll" label and the post button will be enabled even if no activity message is filled.